--- data/mods/gen9godlygift/scripts.ts
@@ -1,4 +1,0 @@
-export const Scripts: ModdedBattleScriptsData = {
-	gen: 9,
-	inherit: 'gen9',
-};
